added better typescript version support
