+++ index.d.ts
@@ -1,4 +1,4 @@
-import type { EventEmitter as EE } from "./index.d.mts";
+import type { EventEmitter as EE } from "./shared-types.d.ts";
 
 declare const EventEmitter: typeof EE;
 export = EventEmitter;
