+++ package.json
@@ -23,6 +23,7 @@
     "index.mjs",
     "index.d.ts",
     "index.d.mts",
+    "shared-types.d.ts",
     "dist"
   ],
   "repository": {
